backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
openai==1.106.1
python-docx==1.2.0
PyPDF2==3.0.1
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
aiofiles==24.1.0
# 版本兼容性（放宽范围给依赖解析）
anyio>=4,<5
# 新增的文档处理库
pdfplumber==0.11.7
pymupdf==1.26.4
docx2python==3.5.0
docxcompose==1.4.0
requests==2.32.3
asyncio-throttle==1.0.2
duckduckgo-search==8.1.1
# LangChain支持（已移除）
beautifulsoup4==4.12.3
# 增强爬虫能力
playwright==1.51.0
seleniumbase==4.33.3
undetected-chromedriver==3.5.5
# MCP服务支持
mcp==1.13.1